fixed warnings

+++ include/runner.hpp
@@ -17,7 +17,7 @@ class Runner
     auto run_instruction(const Instruction& instruction) -> void;
 
     auto check_for_flags(std::uint16_t source, std::uint16_t dest,
-                         const std::function<uint16_t(uint16_t, uint16_t)>& opr) -> void;
+                         const std::function<int32_t(uint16_t, uint16_t)>& opr) -> void;
     auto set_negative() -> void;
     auto set_zero() -> void;
     auto set_overflow() -> void;
